refactored testing suite. general clean up but moved shape inference functions and other helper functions to utils.py so they can be reused easily across tests. returned testing stucture to the setup where each file is for a seperate op.
